# Conflicts:
#	time-capsule-service-logic/src/main/java/cz/cvut/fel/pm2/api/UserApiImpl.java

+++ time-capsule-service-logic/src/main/java/cz/cvut/fel/pm2/repository/CapsuleRepository.java
@@ -32,7 +32,7 @@ public interface CapsuleRepository extends JpaRepository<Capsule, Long> {
      * @param id id of the capsule
      * @return capsule with the given id
      */
-    Optional<Capsule> getCapsuleById(Integer id);
+    Optional<Capsule> getCapsuleById(Long id);
 
     /**
      * Find capsule by qr code password
